Refactor: Mv `TarEntriesVisitor` to mod `async_extracter`
